+++ backend/src/main/java/ru/hh/career/solution/area/service/ParserAreaService.java
@@ -1,10 +1,10 @@
 package ru.hh.career.solution.area.service;
 
 import org.springframework.stereotype.Service;
-import ru.hh.career.solution.area.dao.CheckerOnNullDao;
 import ru.hh.career.solution.area.dto.AreaDto;
 import ru.hh.career.solution.area.dto.AreaParser;
 import ru.hh.career.solution.area.entity.Area;
+import ru.hh.career.solution.dao.CheckerOnNullDao;
 import ru.hh.career.solution.dao.GenericDao;
 
 import javax.inject.Inject;
