+++ backend/src/main/java/ru/hh/career/solution/dao/CheckerOnNullDao.java
@@ -1,7 +1,7 @@
-package ru.hh.career.solution.professionalenvironment.dao;
+package ru.hh.career.solution.dao;
 
 import org.hibernate.SessionFactory;
-import ru.hh.career.solution.dao.GenericDao;
+import ru.hh.career.solution.area.entity.Area;
 import ru.hh.career.solution.professionalenvironment.entity.ProfessionalEnvironment;
 
 import javax.inject.Inject;
@@ -20,6 +20,15 @@ public class CheckerOnNullDao extends GenericDao {
     CriteriaBuilder criteriaBuilder = getSession().getCriteriaBuilder();
     CriteriaQuery<Long> query = criteriaBuilder.createQuery(Long.class);
     Root<ProfessionalEnvironment> root = query.from(ProfessionalEnvironment.class);
+    query.select(criteriaBuilder.count(root));
+    long resultCount = getSession().createQuery(query).getSingleResult();
+    return resultCount > 0;
+  }
+
+  public boolean isAreaPresent() {
+    CriteriaBuilder criteriaBuilder = getSession().getCriteriaBuilder();
+    CriteriaQuery<Long> query = criteriaBuilder.createQuery(Long.class);
+    Root<Area> root = query.from(Area.class);
     query.select(criteriaBuilder.count(root));
     long resultCount = getSession().createQuery(query).getSingleResult();
     return resultCount > 0;
